22년 2월 04일 Update가 안된다.. 고장났다.. 고쳐줘라. .

+++ src/main/java/com/example/studytest1/dto/HYTestDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 @Getter
 @Setter
 public class HYTestDto {
-    private int seq;
+    private Long seq;
     private String name;
     private String addr;
     private String phonenumber;
